added OfferCalculationServiceImplTest and optimized CreditCalculationServiceImplTest
